[BE]평균 별점 double로 타입 변환

+++ back/src/main/java/seb40_main_012/back/book/BookDto.java
@@ -52,7 +52,7 @@ public class BookDto {
         private long bookId;
         private Genre genre;
         private BookWiki bookWiki;
-        private long averageRating;
+        private double averageRating;
         private List<Comment> comments;
         private List<Pairing> pairings;
         private List<BookCollection> bookCollections;
