+++ back/src/main/java/seb40_main_012/back/book/entity/Book.java
@@ -44,7 +44,7 @@ public class Book {
     /*도서 별점*/
     //    --------------------------------------------------------------------------------------------
     @Column
-    private long averageRating;
+    private double averageRating;
     @Column
     private long ratingCount;
     //    @JsonManagedReference
